FEAT: assetStore를 생성해서 리팩토링
