+++ src/components/AssetPage/AssetList.vue
@@ -1,111 +1,72 @@
 <template>
   <div class="asset-list">
     <h1 class="asset-list__title">자산 정보</h1>
-    <!-- 자산 목록 페이지의 메인 타이틀 -->
-
-    <!-- 데이터를 불러오는 동안 로딩 스피너 또는 메시지 표시 -->
-    <div v-if="loading" class="asset-list__loading">Loading...</div>
-
-    <!-- 데이터 가져오기에 오류가 있을 때 오류 메시지 표시 -->
-    <div v-else-if="error" class="asset-list__error">{{ error }}</div>
-
-    <!-- 데이터가 성공적으로 처리되었을 때 메인 콘텐츠 영역 표시 -->
-    <template v-else-if="processedData">
+    <div v-if="assetStore.loading" class="asset-list__loading">Loading...</div>
+    <div v-else-if="assetStore.error" class="asset-list__error">{{ assetStore.error }}</div>
+    <template v-else-if="assetStore.processedData">
       <div class="asset-list__grid">
-        <!-- 섹션: 전체 자산 금액을 표시하는 TotalAsset 컴포넌트 -->
         <div class="asset-list__section asset-list__total">
-          <!-- 전체 자산 즉 총 자산을 나타내는 부분 -->
-          <TotalAsset :totalAmount="processedData.totalAsset" />
+          <TotalAsset :totalAmount="assetStore.processedData.totalAsset" />
         </div>
-
-        <!-- 섹션: 자산 분포 및 평균과의 비교를 위한 슬라이더 -->
         <div class="asset-list__section asset-list__distribution-slider">
-          <!-- 이전 슬라이드 버튼 -->
           <button @click="prevSlide" class="slider-btn prev-btn">‹</button>
-
-          <!-- 슬라이더 콘텐츠 영역 -->
           <div class="slider-content">
-            <!-- 페이드 효과로 슬라이드 전환 -->
             <transition name="fade" mode="out-in">
-              <!-- 첫 번째 슬라이드: 사용자 자산 분포를 나타냄 -->
               <Distribution
                 v-if="currentSlide === 0"
                 key="user-distribution"
-                :assetDetails="processedData.assetDetails"
+                :assetDetails="assetStore.processedData.assetDetails"
                 title="내 자산 분포"
               >
-                <!-- 추가 콘텐츠 영역: 가장 많은 자산 종류 표시 -->
                 <template #extra>
                   <div v-if="highestAssetType" class="asset-highlight">
                     보유 자산 중 {{ assetNames[highestAssetType] }}이 제일 많아요!
                   </div>
                 </template>
               </Distribution>
-
-              <!-- 두 번째 슬라이드: 유형별 평균 자산 분포를 나타냄 -->
               <DistributionAverage
-                v-else-if="currentSlide === 1 && processedData.typeAverages"
+                v-else-if="currentSlide === 1 && assetStore.processedData.typeAverages"
                 key="type-average-distribution"
-                :assetDetails="processedData.typeAverages"
-                :title="`${processedData.assetDetails.type || '전체'} 평균 자산 분포`"
+                :assetDetails="assetStore.processedData.typeAverages"
+                :title="`${assetStore.processedData.assetDetails.type || '전체'} 평균 자산 분포`"
               />
-
-              <!-- 세 번째 슬라이드: 전체 사용자 평균 자산 분포 -->
               <DistributionAverage
-                v-else-if="currentSlide === 2 && processedData.overallAverages"
+                v-else-if="currentSlide === 2 && assetStore.processedData.overallAverages"
                 key="overall-average-distribution"
-                :assetDetails="processedData.overallAverages"
+                :assetDetails="assetStore.processedData.overallAverages"
                 title="전체 사용자 평균 자산 분포"
               />
             </transition>
           </div>
-
-          <!-- 다음 슬라이드 버튼 -->
           <button @click="nextSlide" class="slider-btn next-btn">›</button>
         </div>
-
-        <!-- 섹션: 자산 비교 차트 영역 -->
         <div class="asset-list__section asset-list__comparison-container">
-          <!-- 자산 종류 버튼 -->
           <AssetTypeButtons :selectedType="selectedAssetType" @select-type="selectAssetType" />
-
-          <!-- 자산 비교 차트 -->
           <div class="asset-list__comparison-charts">
-            <!-- 사용자 자산과 선택된 자산 종류 간 비교 -->
             <div class="asset-list__chart">
               <AssetComparison
-                :userAsset="calculateTotalAssets(processedData.assetDetails)"
-                :userType="processedData.assetDetails.type || 'unknown'"
+                :userAsset="calculateTotalAssets(assetStore.processedData.assetDetails)"
+                :userType="assetStore.processedData.assetDetails.type || 'unknown'"
                 :selectedAssetType="selectedAssetType"
               />
             </div>
-
-            <!-- 시간에 따른 자산 변화 비교 -->
             <div class="asset-list__chart">
-              <TimeComparison :assetType="selectedAssetType" :assetData="rawAssetData" />
+              <TimeComparison :assetType="selectedAssetType" :assetData="assetStore.rawAssetData" />
             </div>
           </div>
         </div>
-
-        <!-- 섹션: 대출 정보 -->
         <div class="asset-list__section asset-list__loan">
           <h2 class="section-title">대출 정보</h2>
-          <!-- 대출 정보가 있는 경우 LoanInfo 컴포넌트로 대출 정보 표시 -->
           <template v-if="hasLoanData">
-            <LoanInfo :loanData="processedData.loanData" />
+            <LoanInfo :loanData="assetStore.processedData.loanData" />
           </template>
-          <!-- 대출 정보가 없는 경우 안내 메시지 표시 -->
           <p v-else class="no-loan-message">대출이 존재하지 않습니다.</p>
         </div>
-
-        <!-- 섹션: 추천 상품 -->
         <div class="asset-list__section asset-list__recommended-products">
           <h2 class="section-title">추천 상품</h2>
-          <!-- 대출 정보가 있는 경우 대출 기간에 따른 추천 상품 표시 -->
           <template v-if="hasLoanData">
-            <Recommendation :loanPeriod="processedData.loanData.period" />
+            <Recommendation :loanPeriod="assetStore.processedData.loanData.period" />
           </template>
-          <!-- 대출 정보가 없을 경우 안내 메시지 표시 -->
           <p v-else class="no-recommendation-message">대출 정보가 없어 추천 상품이 없습니다.</p>
         </div>
       </div>
@@ -117,7 +78,7 @@
 //src/components/AssetPage/AssetList.vue
 
 import { ref, computed, onMounted } from 'vue'
-import { fetchAssetData, fetchAssetAverages } from '@/api/assetApi'
+import { useAssetStore } from '@/stores/assetStore'
 import TotalAsset from '@/components/AssetPage/TotalAsset.vue'
 import Distribution from '@/components/AssetPage/Distribution.vue'
 import AssetTypeButtons from '@/components/AssetPage/AssetTypeButtons.vue'
@@ -127,16 +88,10 @@ import LoanInfo from '@/components/AssetPage/LoanInfo.vue'
 import Recommendation from '@/components/AssetPage/Recommendation.vue'
 import DistributionAverage from '@/components/AssetPage/DistributionAverage.vue'
 
-const loading = ref(true) // 로딩 상태 관리
-const error = ref(null) // 에러 상태 관리
-
-const rawAssetData = ref(null) // 자산 데이터 원본
-const assetAverages = ref(null) // 평균 자산 데이터
-const selectedAssetType = ref('cash') // 선택된 자산 타입 기본값은 'cash'
-
-//슬라이드 구현
+const assetStore = useAssetStore()
+const selectedAssetType = ref('cash')
 const currentSlide = ref(0)
-const totalSlides = 3 // 전체 슬라이드 수를 3으로 변경
+const totalSlides = 3
 
 const nextSlide = () => {
   currentSlide.value = (currentSlide.value + 1) % totalSlides
@@ -152,49 +107,21 @@ const assetNames = {
   stock: '주식',
   insurance: '보험'
 }
-// 자산 데이터 및 평균 데이터를 API로부터 로드하는 함수
 
-const fieldMapping = {
-  cash: { bank: 'cashBank', account: 'cashAccount', value: 'cash' },
-  deposit: { bank: 'depositBank', account: 'depositAccount', value: 'deposit' },
-  stock: { bank: 'stockBank', account: 'stockAccount', value: 'stock' },
-  insurance: { bank: 'insuranceCompany', account: 'insuranceName', value: 'insurance' }
-}
 const highestAssetType = computed(() => {
-  if (!processedData.value || !processedData.value.assetDetails) return null
-  const sortedAssets = Object.entries(processedData.value.assetDetails).sort(
+  if (!assetStore.processedData || !assetStore.processedData.assetDetails) return null
+  const sortedAssets = Object.entries(assetStore.processedData.assetDetails).sort(
     ([, a], [, b]) => b.total - a.total
   )
   return sortedAssets[0]?.[0]
 })
-const loadData = async () => {
-  console.log('1. loadData 함수 시작')
-  try {
-    loading.value = true
-    const [assetData, averages] = await Promise.all([fetchAssetData(), fetchAssetAverages()])
-    console.log('2. 받은 평균 데이터:', averages)
-    rawAssetData.value = assetData
-    assetAverages.value = averages
-    console.log('3. assetAverages.value 설정됨:', assetAverages.value)
-  } catch (err) {
-    console.error('데이터 가져오기 실패:', err)
-    error.value = '데이터를 불러오는데 실패했습니다. 나중에 다시 시도해주세요.'
-  } finally {
-    loading.value = false
-    console.log('4. loadData 함수 종료')
-  }
-}
 
-// 자산 데이터를 처리하여 필요한 형태로 변환하는 함수
+const hasLoanData = computed(() => {
+  if (!assetStore.processedData || !assetStore.processedData.loanData) return false
+  const { amount, purpose, period, interest } = assetStore.processedData.loanData
+  return amount > 0 || purpose || period > 0 || interest > 0
+})
 
-const parseJsonArray = (jsonString) => {
-  try {
-    return JSON.parse(jsonString)
-  } catch (e) {
-    console.error('Failed to parse JSON string:', jsonString)
-    return []
-  }
-}
 const calculateTotalAssets = (assetDetails) => {
   const totals = {}
   for (const [type, data] of Object.entries(assetDetails)) {
@@ -202,143 +129,13 @@ const calculateTotalAssets = (assetDetails) => {
   }
   return totals
 }
-// 자산 데이터를 처리하여 필요한 형태로 변환하는 함수
 
-//따로 함수 처리 ㅎ
-
-const processAssetData = (data, assetTypes) => {
-  return assetTypes.reduce((processed, type) => {
-    const mapping = fieldMapping[type]
-    if (mapping) {
-      const values = parseJsonArray(data[mapping.value])
-      const banks = parseJsonArray(data[mapping.bank])
-      const accounts = parseJsonArray(data[mapping.account])
-
-      processed[type] = {
-        values: values.length ? values.map((v) => Number(v) || 0) : [0],
-        banks: banks.length ? banks : [''],
-        accounts: accounts.length ? accounts : ['']
-      }
-    }
-    return processed
-  }, {})
-}
-const calculateAverage = (values) => {
-  const validValues = values.filter((v) => v > 0)
-  return validValues.length
-    ? validValues.reduce((sum, val) => sum + val, 0) / validValues.length
-    : 0
-}
-
-// 처리된 자산 데이터를 계산하고 반환하는 computed 함수
-const processedData = computed(() => {
-  if (!rawAssetData.value || rawAssetData.value.length === 0) {
-    return {
-      totalAsset: 0,
-      assetDetails: {
-        cash: { total: 0, details: [] },
-        deposit: { total: 0, details: [] },
-        stock: { total: 0, details: [] },
-        insurance: { total: 0, details: [] },
-        type: '유형을 입력해 주세요'
-      },
-      comparisonData: {},
-      typeAverages: null,
-      timeComparisonData: {},
-      loanData: { amount: 0, purpose: '', period: 0, interest: 0 },
-      overallAverages: null
-    }
-  }
-
-  const currentData = rawAssetData.value[0] || {}
-  const previousData = rawAssetData.value[1] || currentData
-  const assetTypes = ['cash', 'deposit', 'stock', 'insurance']
-
-  const currentAssetData = processAssetData(currentData, assetTypes)
-  const previousAssetData = processAssetData(previousData, assetTypes)
-
-  const calculateTotal = (assetData) =>
-    Object.values(assetData).reduce(
-      (total, { values }) => total + values.reduce((sum, val) => sum + Number(val), 0),
-      0
-    )
-
-  const totalAsset = calculateTotal(currentAssetData)
-
-  const assetDetails = assetTypes.reduce((details, type) => {
-    const values = currentAssetData[type].values
-    details[type] = {
-      total: values.reduce((sum, val) => sum + val, 0),
-      details: values.map((value, index) => ({
-        bank: currentAssetData[type].banks[index] || '',
-        account: currentAssetData[type].accounts[index] || '',
-        value: value
-      }))
-    }
-    return details
-  }, {})
-
-  const comparisonData = assetTypes.reduce((data, type) => {
-    const currentValues = currentAssetData[type].values
-    const previousValues = previousAssetData[type].values
-    data[type] = {
-      average: assetAverages.value ? assetAverages.value[type] : 0,
-      user: calculateAverage(currentValues),
-      previousAverage: calculateAverage(previousValues)
-    }
-    return data
-  }, {})
-
-  const timeComparisonData = assetTypes.reduce((data, type) => {
-    data[type] = {
-      previousMonth: calculateAverage(previousAssetData[type].values),
-      currentMonth: calculateAverage(currentAssetData[type].values)
-    }
-    return data
-  }, {})
-
-  const typeAverages = assetAverages.value
-    ? assetTypes.reduce((averages, type) => {
-        averages[type] = { total: assetAverages.value[type] || 0 }
-        return averages
-      }, {})
-    : null
-  // 새로 추가된 부분: overallAverages 계산
-  const overallAverages = assetAverages.value
-    ? assetTypes.reduce((averages, type) => {
-        averages[type] = { total: assetAverages.value[type] || 0 }
-        return averages
-      }, {})
-    : null
-
-  return {
-    totalAsset,
-    assetDetails: { ...assetDetails, type: currentData.type || '유형을 입력해주세요' },
-    comparisonData,
-    typeAverages,
-    overallAverages, // 여기에 overallAverages 추가
-    timeComparisonData,
-    loanData: {
-      amount: Number(currentData.loanAmount),
-      purpose: currentData.loanPurpose,
-      period: Number(currentData.period),
-      interest: Number(currentData.interest)
-    }
-  }
-})
-
-// 대출 정보 존재 여부를 확인하는 computed 속성
-const hasLoanData = computed(() => {
-  if (!processedData.value || !processedData.value.loanData) return false
-  const { amount, purpose, period, interest } = processedData.value.loanData
-  return amount > 0 || purpose || period > 0 || interest > 0
-})
 const selectAssetType = (type) => {
   selectedAssetType.value = type
 }
 
-onMounted(async () => {
-  await loadData()
+onMounted(() => {
+  assetStore.loadData()
 })
 </script>
