+++ src/pages/asset/AssetPage.vue
@@ -1,7 +1,6 @@
 <template>
   <div class="asset-page">
     <AssetList />
-    <AssetTypeDistribution />
   </div>
 </template>
